refactor: 폴더 정리 및 prefetch 적용
