--- apps/client/src/app/[user]/create-group/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-import CreateGroup from "@/app/[user]/create-group/components/CreateGroup";
-
-const CreateGroupPage = () => {
-  return <CreateGroup />;
-};
-
-export default CreateGroupPage;
