+++ apps/client/src/app/group/[groupId]/@modal/(.)notice/[noticeId]/page.tsx
@@ -1,18 +1,18 @@
 "use client";
 
 import { useNoticeByIdQueryObject } from "@/app/api/notices/query";
-import NoticeDetail from "@/app/group/[groupId]/components/NoticeModal/NoticeDetail";
+import NoticeDetail from "@/app/group/[groupId]/@modal/(.)notice/components/NoticeModal/NoticeDetail";
 import {
   noticeHeaderStyle,
   noticeModalWrapper,
-} from "@/app/group/[groupId]/components/NoticeModal/index.css";
+} from "@/app/group/[groupId]/@modal/(.)notice/components/NoticeModal/index.css";
 import { textStyle } from "@/app/group/[groupId]/components/index.css";
 import Spinner from "@/common/component/Spinner";
 import { usePvEvent } from "@/shared/hook/usePvEvent";
 import { alignCenterStyle } from "@/styles/shared.css";
 import { useQuery } from "@tanstack/react-query";
 
-const NoticeDetailPage = ({
+const InterceptingNoticeDetailPage = ({
   params: { noticeId, groupId },
 }: { params: { groupId: string; noticeId: string } }) => {
   usePvEvent("group_notice_detail_page_view", {
@@ -34,4 +34,4 @@ const NoticeDetailPage = ({
   );
 };
 
-export default NoticeDetailPage;
+export default InterceptingNoticeDetailPage;
