+++ apps/client/src/app/group/[groupId]/@modal/(.)notice/components/NoticeModal/NoticeCreate/schema.ts
