+++ apps/client/src/app/group/[groupId]/@modal/(.)notice/components/NoticeModal/NoticeDetail/index.css.ts
