+++ apps/client/src/app/group/[groupId]/@modal/(.)notice/components/NoticeModal/NoticeList/index.css.ts
