+++ apps/client/src/shared/component/CommentBox/hook.ts
@@ -3,7 +3,7 @@
 import { CommentsContext } from "@/app/@modal/(...)group/[groupId]/solved-detail/components/CommentSection/provider";
 import { useEditCommentMutation } from "@/app/api/comments/mutation";
 import { usePatchNoticeCommentMutation } from "@/app/api/notices/mutation";
-import { NoticeCommentsContext } from "@/app/group/[groupId]/components/NoticeModal/NoticeDetail/provider";
+import { NoticeCommentsContext } from "@/app/group/[groupId]/@modal/(.)notice/components/NoticeModal/NoticeDetail/provider";
 import { type KeyboardEvent, useContext } from "react";
 import { flushSync } from "react-dom";
 import { type SubmitHandler, useForm } from "react-hook-form";
